Users/rajes/OneDrive/Desktop/claude filter/AutoFilterBot/requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.2
aiohttp==3.9.3
python-dotenv==1.0.1
